// src/features/marketing-blocks/schemas/marketing-blocks.ts
export const statItemSchema = z.object({
  id: z.string(),
  value: z.coerce.number().default(0),
  prefix: z.string().default(""),
  suffix: z.string().default(""),
  labelEn: z.string().default(""),
  labelAr: z.string().default(""),
  descriptionEn: z.string().default(""),
  descriptionAr: z.string().default(""),
  icon: z.string().default(""),
  chartType: z.enum(["none", "bar", "donut"]).default("none"),
  chartData: z.array(z.coerce.number()).default([]),
});
// ...
export type StatItem = z.infer<typeof statItemSchema>;